fix(pro): share abandoned cart terminal state resolution
Centralize workflow terminal-state discovery so scheduler and job processing use the same cached logic and do not drift when state-machine APIs change.

Made-with: Cursor
